requirements.txt
fastapi
uvicorn
tensorflow
tensorflow-hub
pillow
